Day 1 work
using MyClass.h I was able to create a phaser object and attach its parameters to my sliders. Now I am working on using ChorusPhaserProcessor.h to have a UI where I choose the effect on the GUI and the effect specific knobs only show up and manipulate the sound from there. I got both header files from ChatGPT and cleaned them up and not entirely done with making the DSP work. After that is done, implement the combobox and create an if function that allows the effect specific parameters get the AddAndMakeVisible Command.